+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/iamv2/WorkspaceAccessDetailAccessType.java
@@ -1,0 +1,12 @@
+// Code generated from OpenAPI specs by Databricks SDK Generator. DO NOT EDIT.
+
+package com.databricks.sdk.service.iamv2;
+
+import com.databricks.sdk.support.Generated;
+
+/** The type of access the principal has to the workspace. */
+@Generated
+public enum WorkspaceAccessDetailAccessType {
+  DIRECT,
+  INDIRECT,
+}