+++ databricks-sdk-java/src/main/java/com/databricks/sdk/service/jobs/RunTask.java
@@ -84,10 +84,6 @@ public class RunTask {
   @JsonProperty("description")
   private String description;
 
-  /** Deprecated, field was never used in production. */
-  @JsonProperty("disabled")
-  private Boolean disabled;
-
   /**
    * The actual performance target used by the serverless run during execution. This can differ from
    * the client-set performance target on the request depending on whether the performance mode is
@@ -417,15 +413,6 @@ public class RunTask {
 
   public String getDescription() {
     return description;
-  }
-
-  public RunTask setDisabled(Boolean disabled) {
-    this.disabled = disabled;
-    return this;
-  }
-
-  public Boolean getDisabled() {
-    return disabled;
   }
 
   public RunTask setEffectivePerformanceTarget(PerformanceTarget effectivePerformanceTarget) {
@@ -759,7 +746,6 @@ public class RunTask {
         && Objects.equals(dbtTask, that.dbtTask)
         && Objects.equals(dependsOn, that.dependsOn)
         && Objects.equals(description, that.description)
-        && Objects.equals(disabled, that.disabled)
         && Objects.equals(effectivePerformanceTarget, that.effectivePerformanceTarget)
         && Objects.equals(emailNotifications, that.emailNotifications)
         && Objects.equals(endTime, that.endTime)
@@ -811,7 +797,6 @@ public class RunTask {
         dbtTask,
         dependsOn,
         description,
-        disabled,
         effectivePerformanceTarget,
         emailNotifications,
         endTime,
@@ -863,7 +848,6 @@ public class RunTask {
         .add("dbtTask", dbtTask)
         .add("dependsOn", dependsOn)
         .add("description", description)
-        .add("disabled", disabled)
         .add("effectivePerformanceTarget", effectivePerformanceTarget)
         .add("emailNotifications", emailNotifications)
         .add("endTime", endTime)